--- src/main/java/com/CloudClub/jpaStudy/controller/MemberForm.java
@@ -1,16 +1,0 @@
-package com.CloudClub.jpaStudy.controller;
-
-import jakarta.validation.constraints.NotEmpty;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter @Setter
-public class MemberForm {
-
-	@NotEmpty(message = "이름은 필수")
-	private String name;
-
-	private String city;
-	private String street;
-	private String zipcode;
-}
