+++ src/main/java/com/CloudClub/jpaStudy/controller/member/dtos/MemberForm.java
@@ -1,0 +1,17 @@
+package com.CloudClub.jpaStudy.controller.member.dtos;
+
+import jakarta.validation.constraints.NotEmpty;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class MemberForm {
+
+  @NotEmpty(message = "이름은 필수")
+  private String name;
+
+  private String city;
+  private String street;
+  private String zipcode;
+}
